Add batch file
The `build.bat` file builds both the 32-bit and 64-bit versions of the
DLL on Windows. The assumption is that it would be run from a x64 Native
Tools command prompt. This reduces the number of commands to use to
build on Windows to just one, i.e. `> build.bat`.