Fix install/build artifacts location for labssh2
There was a redundant `bin\bin` folder for the output of the build. This
was caused by the "install" directory being the `bin` folder of the
current CMake binary directory (`CMAKE_CURRENT_BINARY_DIR`). I wanted
the output of the entire build to be easily located and only contain the
DLL. Thus, I moved the "install" location for everything from labssh2 to
the `labssh2` sub-build folder and the install binary location to the
`bin` folder. This removed the redundant folders and placed the other
build/install output from the labssh2 project out of the way in the
`labssh2` sub-folder of the build.